pme2: add pme2_have_control(), and fix non-ctrl plane linux
pme2_have_control() checks at run-time to see if ccsr space is
available. Also changed API for pme_attr_get_attr() to return error.

Updated scan test to not run on non-ctrl plane (it depends on sample
db to be present, which may not be the case if sample db is still
running on the control plane).

Signed-off-by: Geoff Thorpe <geoff.thorpe@freescale.com>
Signed-off-by: Jeffrey Ladouceur <Jeffrey.Ladouceur@freescale.com>