dpa: Add support for zero-copy forLinux-only transactions
Linux does not have a way of returning buffers to arbitrary software
drivers when it is done with them, so we need to allocate new buffers to
replace the ones we give to the stack.  This is the mechanism:

Each pool maintains an array of pointers to the pages in which these
buffers exist.  For each buffer in the pool, there is a page pointer in
the array, which is used to remember the mapping between virtual and
physical addresses for the page.  When a buffer is received, its
corresponding page pointer in the array is found, and used to link up
the skb, and unmap the buffer (wrt to the DPA enet).  The page struct's
_count field is appropriately incremented such that the page will be
freed when the kernel is done with *all* buffers contained therein.

Signed-off-by: Andy Fleming <afleming@freescale.com>