p4080sim: revert DTS and defconfig to previous versions
Commit "Add PHY support for working RGMII/SGMII" modified the P4080SIM
device tree and defconfig, instead of the P4080DS version of these files.
This is because the P4080DS version did not exist at the time.  However, we
now have P4080DS versions of these files, so we should revert the P4080SIM
versions.