Add support for sharing a MAC across partitions
In order to share a MAC, we add support for specifying which frame
queues to use even for MAC-connected ethernet nodes.  This allows
all partitions to share the same TX queues so they can all transmit
through the MAC.  For RX to work, fmc needs to be run with appropriate
settings.

In the process, I cleaned up a few things in the neighborhood, and
started moving some of this code toward refactoring.  Also, removed
some driver-specific conditions around error prints so that errors
are reported always.

Signed-off-by: Andy Fleming <afleming@freescale.com>