Add support for the p4080ds_mdio "device"
The P4080 uses a peculiar muxing scheme where the two MDIO buses pinned
out from the p4080 are redirected to one of 6 (4 for 1G, 2 for 10G) bus
lines on the DS.  This requires that for each MDIO operation, we write
to some GPIO pins to talk to the desired PHY.  However, the PHY Lib and
MDIO driver are unaware of this, and should remain unaware of this if we
are to cleanly run p4080 code on other boards (and on other chips which
use the same IP).

As such, we design a "fake" MDIO bus type which is specific to the
P4080DS, and use it as a wrapper around the actual bus(es).  This
requires some device tree changes:

1) Addition of p4080ds mdio nodes
2) Moving all PHYs under those nodes
3) Adding aliases and labels so that u-boot can find those nodes in a
less fragile way.
4) Correction of the broken MDIO nodes

While we're at it, we also add full TBI support to the ethernet nodes,
so that the dTSEC can use the same code for MDIO

Also, we need to probe for the p4080ds mdio devices in the board code.

Signed-off-by: Andy Fleming <afleming@freescale.com>