Support ethernet on Linux without MDIO
In order for ethernet to work when running under a hypervisor, the
OS without the MDIO needs some means of ascertaining the link properties.
This patch does not fully solve this problem, but allows for the
workaround to use a statically-declared link state, so that the dpa driver
can still operate using PHY Lib.

Thus:
* Fix some line-length issues
* Add support for using dpa with a fixed-link PHY

Signed-off-by: Andy Fleming <afleming@freescale.com>