dpa: Fix VLAN support
When using DPA with a VLAN, the reserved space was being partially used
for the VLAN tag, which meant that we no longer had DPA_BP_HEAD
available.  However, we only really needed 4 bytes for the skb pointer,
so rather than do some very complicated math, we just back up to the
beginning of the buffer, and put the skb pointer there.  (If there isn't
even room for the skb pointer, we call skb_realloc_headroom()).

Signed-off-by: Andy Fleming <afleming@freescale.com>