hvc-fsl: fix dropping of characters when output byte channel is full
The Freescale Hypervisor byte channel console driver aborts the write
operation if the hypervisor returns an error during a write to a byte channel.
This included the error EAGAIN, which only means that the byte channel is full.
When the driver receives EAGAIN, it tells hvc how many characters it did
transmit.  Unfortunately, if this value is 0, hvc assumes that an error occured,
and so it aborts the current write as well.

Modify the driver to return a negative error code if an error other than
EAGAIN occurs.

Enable the IRQ support, which somehow never got fully coded.

Finally, modify hvc itself to consider 0 not an error, but instead just loop
and call the client driver (e.g. hvc_fsl) again.  This will continue until the
client driver returns a real error or sends at least one byte to the byte
channel, or until a timeout occurs.

Signed-off-by: Timur Tabi <timur@freescale.com>