e500mc: enable FP emulation to work around errata CPU12 and CPU13
Errata CPU12 and CPU13 mean that floating-point operations are broken, so we
need to enable FP emulation in the kernel.  However, e500mc has normal FP
support, so we need to modify the kernel to call the software FP emulation
handler instead of the normal FP handler.  To do this, we define a new
exception entry point called FloatingPointUnavailableErrata and set IVOR7 to
it.

This hack is not compatible with rev2 silicon.

Signed-off-by: Timur Tabi <timur@freescale.com>