P4080/PM: Add an idle loop based on the wait instruction, and use it on e500mc.
e500mc cannot doze due to an erratum, but it has a "wait" instruction
that is similar.  It is implemented as a separate idle loop rather than
an inner-loop power_save() callback, because we don't want IRQs or
IRQ-off tracing disabled, and most of the rest of cpu_idle() is
unnecessary even if possibly harmless.

Note that on platforms that select this idle loop will not be able to use
nap-on-idle.

Signed-off-by: Scott Wood <scottwood@freescale.com>