pme2: flush PME onboard cache before deallocating flow contexts.
This necessitates a change to the pme_ctx_disable() API, which will now
kick off an async update-flow to flush the cache. In this case,
pme_ctx_disable() returns +1, indicating that the disable completes once
the provided token has its callback invoked. For direct-mode contexts,
the return value is 0 for success, the token is unused.

This also fixes a significant bug in pme_ctx_disable() that would have
prevented a disabled context from ever being re-enabled properly (the
write-enable mask for restoring the tx FQD fields was incorrect).

This also fixes PMTCC contexts by ensuring they are also DIRECT and
EXCLUSIVE.

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>
Signed-off-by: Jeffrey Ladouceur <Jeffrey.Ladouceur@freescale.com>