qman: update FD structure for rev2
The "pid" field is replaced with "liodn_offset" and the high-bits of the
address are reduced to 8-bit to make space for "eliodn_offset", all in
keeping with rev2 definitions. This should be backwards compatible with
rev1 code given that "pid" has no meaningful caller use and addresses with
more than 40 bits are pointless too.

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>