qman: run-time workarounds for rev1 errata
Detects Qman IP revision at run-time and keys errata workarounds off it. If CCSR
is available (control-plane) then this determines the revision level otherwise
the compatible property of portal nodes is checked for a revision-specific
string, finaly defaulting to rev1 (with a logged warning). If CCSR is available
and portal nodes also specify the revision, these are compared to ensure they
agree.

This also corrects the following;
* the need to initialise CGR is not an erratum, it's defined behaviour for all
  versions going forward.
* copyrights updated to 2010.
* old cycle-accurate limitations no longer apply (and were never errata, so
  remove the options to eliminate rev1/rev2 confusion).

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>
Signed-off-by: Jeffrey Ladouceur <Jeffrey.Ladouceur@freescale.com>
Signed-off-by: Paul Barrette <paul.barrette@freescale.com>