crypto: caam - algapi: convert cipher key to inlined in shared descriptor
Change key specification in shared descriptor to inline the cipher key.
the protected ('covered') authentication key can't be immediate/inlined,
even if it would fit in a descriptor (CAAM_MAX_DESC_SIZE).

This results in faster performance since the h/w doesn't have to make
the additional access to memory to obtain the cipher key.  A consistent
throughput improvement of 8kpps was observed in a 64-byte packet test.

Signed-off-by: Kim Phillips <kim.phillips@freescale.com>
Signed-off-by: Kumar Gala <galak@kernel.crashing.org>