dpa: Prevent tx backlog from growing unboundedly
When servicing more than one flow of packets, it is possible for one
core to fall behind, which opens up the opportunity for the other core
to overwhelm it with tx confirmation work.  This backlog will grow
unbounded, until we run out of memory.  If you are using pool channels,
this isn't as much of an issue, as one of the other cores will pick up
the slack, but if you make each ethernet device core-affined, then this
happens quickly.

In order to fix this, we use what were previously debug counters to keep
track of how many frames are currently on the tx confirmation software
queue, and if the number grows beyond some arbitrary value (here, we've
chosen 512), then we halt the tx queues.  As this is now a possibilty,
we now also set net_device's trans_start field to "jiffies" whenever we
successfully transmit a packet.

Signed-off-by: Andy Fleming <afleming@freescale.com>