dpa: Move to a tasklet for packet processing
The network rx code is invoked inside softirq context.  If we are not in
softirq context, we have to switch to softirq context before the stack
can process each packet.  This is costly.  Instead, we switch from using
a workqueue to a tasklet, which runs in softirq context.  This allows us
to call netif_receive_skb() directly.

Also, clean up bman_release flags.  Tasklets are not allowed to sleep, so we
can't wait, and for the initial bman_release calls, we need to add
BMAN_RELEASE_FLAG_WAIT so that it actually *does* wait.

Signed-off-by: Andy Fleming <afleming@freescale.com>