dpa: Stop double-freeing skbs on error
The scatter-gather error code was freeing the skb, and then _dpa_rx was
noting the error, and freeing it again.  Also, _dpa_rx was freeing the
skb if netif_rx_ni returned an error, but netif_rx_ni frees the skb if
an error occurs. Lastly, if we're a pool we copy from, we don't want to
return the buffer to the pool, as it's already been returned.

Signed-off-by: Andy Fleming <afleming@freescale.com>