dpa: Eliminate zero-copy support
Tests indicate it performs worse than just copying the buffers into
skbs, due to extensive locking needed for the hash table.  Because this
breaks scatter-gather support, we also eliminate that, and modify the
buffer pools so we only have one pool with buffers large enough to hold
any ethernet packet.

Signed-off-by: Andy Fleming <afleming@freescale.com>