dpa: Make the TX queues core-affined
The current code just sets the queue based on a hash of the packet, and
those queues are set to different priorities.  However, there was no
real association between queue and priority, and there was a high chance
that multiple cores would then be attempting to enqueue into the same tx
queue.  Instead, we modify the driver to select the FQ based on the
processor we are on.  This will eliminate all lock contention.  We can
always modify it to do both at some future point if we choose to be more
clever with skb->priority, or to use a more interesting queueing
discipline.

Signed-off-by: Andy Fleming <afleming@freescale.com>