qman: make it easier to use static initialisation.
Because of unions in "struct qm_fd", static initialisation is very

difficult. This change adds untyped versions of the FD format, moves the
"opaque" field to the top of the union (untagged static initialisation of
a union assigns values to the first field of a union), and adds two macros
for declaring FDs.

A similar change is made to allow static initialisation of "struct qm_fqd"
fields for tail-drop, which is convenient for use with qman_init_fq().
This is achieved by removing the opaque version of td_thresh from the
union, which makes sense anyway as it might otherwise be misinterpreted as
a threshold value which it is *not* (it's a exp+mantissa representation).

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>