qman: fix atomic-ctxt problems in hotpotato test.
The problems were two-fold, we were using on_each_cpu() which runs functions
using doorbell interupts, and we were using get_cpu_var()/put_cpu_var() around
all the code. The latter is not necessary because the code is already running
core-affine, and the former is replaced with a home-brewed mechanism.

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>