qman: fix bug in rev1 errata handling.
Due to a rev1 erratum, "Initialize and Schedule FQ" commands are split into two
commands on rev1. If the first command fails (eg. if the FQ is not OOS or
parked), the error handling logic prematurely inverts the management command
tracking state, leading to an infinite loop checking the wrong result register.

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>