crypto: caam - algapi: convert tunnel mode to a special case of transport mode
Previously, it was thought that in order to set the tunnel/transport mode bit
in the shared descriptor, we would need to postpone building the shared
descriptor until we received the encrypt/decrypt request for the first packet
of the session.  We did this by looking up the xfrm_input_state of the first skb,
setting the bit, and continued making subsequent per-packet decisions according
to the mode.

In discussion with Herbert Xu, the linux-crypto maintainer, it was made
clear that this would not be acceptable (for crypto drivers to assume their
input was an skb), and that changes in the driver should be made such
that tunnel mode be a special case of transport mode.

This is what this patch does:

o we now never set the tunnel mode bit in the shared descriptor,

o on encrypt, we always assign next-header on a per-packet basis, and
subsequently always override the shared descriptor's setting using the DPOVRD
command

o on decrypt, we now always ask the h/w to provide its output, i.e, including
outer headers

o we now implement aead_authenc_encrypt - the non IV-generating encrypt(),
which IPSec parallelization pcrypt utilizes, since it does s/w IV chaining on
per-cpu basis.

o it takes us one step closer to preparing the driver for upstream submission

Signed-off-by: Kim Phillips <kim.phillips@freescale.com>