p4080ds: create a single device tree node for both CPCs
Merge the two CPC device tree nodes into a single node that lists both device
register ranges in the reg property.  Update the L2 nodes to point to this
single node.

The P4080 has two CPC (L3 cache) devices, but they act as one.  Instead of
having two CPC nodes in the device tree, we should have a single node that
represents both CPCs.  The reg property of this node has two register ranges
in it.  Code can determine the number of CPCs by counting the number of entries
in the reg property.  The cache-sets and cache-size properties reflect the
combined values of all CPC nodes together.  That is, if there are two 1MB CPC
devices, then the cache-size property will contain 0x200000 (2MB).

Also add the "cache" string to the compatible property, because the code in
U-Boot which updates the compatible property for L3 nodes will be deleted.
All it does is add the "cache" string, so it makes more sense to have the DTS
be correct.

Signed-off-by: Timur Tabi <timur@freescale.com>
Signed-off-by: Roy Zang <tie-fei.zang@freescale.com>