dpa: Free bman buffers if we allocated them
The dpa_bp_free() function was not removing the buffers from the pool,
which meant that an error or a soft-reboot would leave them there.  This
would cause issues, and is wrong, anyway.  However, we don't want to
clean out the pool if the bman driver was the one that created it, so we
resurrect the kernel_pool field, and use that to determine whether we're
responsible for releasing the buffers or not.

Also, use this to determine if we should free the pages, as we should
only do that if we actually allocated them.  Same for dma_unmap.

Signed-off-by: Andy Fleming <afleming@freescale.com>