fman: fix Outstanding DMA settings for 10G ports
When running tests in the lab and the field that involve 10G ports it has been
observed that the default setting for the number of Outstanding DMAs for 10G
ports in FMD is incorrect. With the default settings LWE pktwire and Linux user
space PoC applications experience a steady stream of Rx Errors at traffic rates
around 7-8 Gbps and higher. This is seen both with Rev. 1 and Rev. 2 silicon.

For Rev. 1 silicon, the tests were run at the following clock rates
(core/platform/fman): 1500/700/600 and at 1200/600/450; for Rev. 2 silicon, the
tests were run at 1200/600/450.

In all cases, updating the Outstanding DMAs from 4 to 8 for 10G ports removed
the steady Rx Error occurrence. The patch applied locally to make this change
in FMD is shown below.

R2.1.1 software is expected to be run by customers on Rev 2. silicon, where the
10G ports are required to be used for bench-marking purposes. Fixing this
default to avoid Rx Errors is critical to avoid customer issues with the
release.

Signed-off-by: Mustafa Hamid <R01337@freescale.com>
Signed-off-by: Shlomi Gridish <RM96313@freescale.com>